+++ insh-db/src/lib.rs
@@ -9,6 +9,7 @@ binaries are embedded in the executable, so PostgreSQL does not have to be insta
 
 mod database;
 mod db_conn_pool_events;
+pub mod find_history;
 mod helpers;
 mod schema;
 pub mod search_history;